refactor(index.jsx): extract component
- Counter 앱을 어떻게 그릴지 관리하는 별도 component 분리
- 상태도 분리되 컴포넌트에서 관리하도록 리팩토링
